+++ http_testing/_record.py
@@ -1,0 +1,9 @@
+from attrs import define
+from httpx import Request, Response
+
+
+@define
+class RecordData:
+    title: str
+    request: Request
+    response: Response