+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 0.8.0 (Unreleased)
 - chore: update dependencies
 - feat: improve assert report
+- feat: add attribute history on PageChecker
 
 ## 0.7.0
 - feat: possible to pass string when check content